tech(api): migrate /api/users/{userId}/campaigns/{campaignId}/profile from lib to src
