+++ api/lib/application/users/index.js
@@ -276,31 +276,6 @@ const register = async function (server) {
     },
     {
       method: 'GET',
-      path: '/api/users/{userId}/campaigns/{campaignId}/profile',
-      config: {
-        validate: {
-          params: Joi.object({
-            userId: identifiersType.userId,
-            campaignId: identifiersType.campaignId,
-          }),
-        },
-        pre: [
-          {
-            method: securityPreHandlers.checkRequestedUserIsAuthenticatedUser,
-            assign: 'requestedUserIsAuthenticatedUser',
-          },
-        ],
-        handler: userController.getUserProfileSharedForCampaign,
-        notes: [
-          '- **Cette route est restreinte aux utilisateurs authentifiés**\n' +
-            '- Récupération du profil d’un utilisateur partagé (**userId**) pour la campagne donnée (**campaignId**)\n' +
-            '- L’id demandé doit correspondre à celui de l’utilisateur authentifié',
-        ],
-        tags: ['api', 'user', 'campaign'],
-      },
-    },
-    {
-      method: 'GET',
       path: '/api/users/{userId}/campaigns/{campaignId}/assessment-result',
       config: {
         validate: {
