+++ api/src/prescription/campaign-participation/application/learner-participation-controller.js
@@ -1,8 +1,10 @@
 import { DomainTransaction } from '../../../../lib/infrastructure/DomainTransaction.js';
 import { monitoringTools } from '../../../shared/infrastructure/monitoring-tools.js';
+import * as requestResponseUtils from '../../../shared/infrastructure/utils/request-response-utils.js';
 import { ApplicationTransaction } from '../../shared/infrastructure/ApplicationTransaction.js';
 import { usecases } from '../domain/usecases/index.js';
 import * as campaignParticipationSerializer from '../infrastructure/serializers/jsonapi/campaign-participation-serializer.js';
+import * as sharedProfileForCampaignSerializer from '../infrastructure/serializers/jsonapi/shared-profile-for-campaign-serializer.js';
 
 const save = async function (request, h, dependencies = { campaignParticipationSerializer, monitoringTools }) {
   const userId = request.auth.credentials.userId;
@@ -42,10 +44,32 @@ const beginImprovement = async function (request) {
   });
 };
 
+const getUserProfileSharedForCampaign = async function (
+  request,
+  h,
+  dependencies = {
+    sharedProfileForCampaignSerializer,
+    requestResponseUtils,
+  },
+) {
+  const authenticatedUserId = request.auth.credentials.userId;
+  const campaignId = request.params.campaignId;
+  const locale = dependencies.requestResponseUtils.extractLocaleFromRequest(request);
+
+  const sharedProfileForCampaign = await usecases.getUserProfileSharedForCampaign({
+    userId: authenticatedUserId,
+    campaignId,
+    locale,
+  });
+
+  return dependencies.sharedProfileForCampaignSerializer.serialize(sharedProfileForCampaign);
+};
+
 const learnerParticipationController = {
   save,
   shareCampaignResult,
   beginImprovement,
+  getUserProfileSharedForCampaign,
 };
 
 export { learnerParticipationController };
