+++ api/src/prescription/campaign-participation/domain/usecases/get-user-profile-shared-for-campaign.js
@@ -1,6 +1,6 @@
-import { constants } from '../../../src/shared/domain/constants.js';
-import { NoCampaignParticipationForUserAndCampaign } from '../../../src/shared/domain/errors.js';
-import { SharedProfileForCampaign } from '../../../src/shared/domain/read-models/SharedProfileForCampaign.js';
+import { constants } from '../../../../shared/domain/constants.js';
+import { NoCampaignParticipationForUserAndCampaign } from '../../../../shared/domain/errors.js';
+import { SharedProfileForCampaign } from '../../../../shared/domain/read-models/SharedProfileForCampaign.js';
 
 const getUserProfileSharedForCampaign = async function ({
   userId,
