+++ api/tests/integration/infrastructure/repositories/campaign-participation-repository_test.js
@@ -1,9 +1,7 @@
 import { DomainTransaction } from '../../../../lib/infrastructure/DomainTransaction.js';
 import * as campaignParticipationRepository from '../../../../lib/infrastructure/repositories/campaign-participation-repository.js';
-import { CampaignParticipation } from '../../../../src/prescription/campaign-participation/domain/models/CampaignParticipation.js';
 import { CampaignParticipationStatuses, CampaignTypes } from '../../../../src/prescription/shared/domain/constants.js';
 import { constants } from '../../../../src/shared/domain/constants.js';
-import { Assessment } from '../../../../src/shared/domain/models/Assessment.js';
 import { Campaign } from '../../../../src/shared/domain/models/Campaign.js';
 import { databaseBuilder, expect, sinon } from '../../../test-helper.js';
 
@@ -393,93 +391,6 @@ describe('Integration | Repository | Campaign Participation', function () {
       expect(latestCampaignParticipation1.assessments).to.be.instanceOf(Array);
       expect(latestCampaignParticipation1.campaign).to.be.instanceOf(Campaign);
       expect(latestCampaignParticipations).to.have.lengthOf(2);
-    });
-  });
-
-  describe('#findOneByCampaignIdAndUserId', function () {
-    let userId;
-    let campaignId;
-
-    beforeEach(async function () {
-      userId = databaseBuilder.factory.buildUser().id;
-      const otherUserId = databaseBuilder.factory.buildUser().id;
-
-      campaignId = databaseBuilder.factory.buildCampaign().id;
-      const otherCampaignId = databaseBuilder.factory.buildCampaign().id;
-
-      databaseBuilder.factory.buildCampaignParticipation({
-        campaignId,
-        userId: otherUserId,
-      });
-      databaseBuilder.factory.buildCampaignParticipation({
-        campaignId: otherCampaignId,
-        userId,
-      });
-      await databaseBuilder.commit();
-    });
-
-    it('should return the campaign participation found', async function () {
-      // given
-      const campaignParticipation = databaseBuilder.factory.buildCampaignParticipation({
-        campaignId,
-        userId,
-      });
-      await databaseBuilder.commit();
-
-      // when
-      const response = await campaignParticipationRepository.findOneByCampaignIdAndUserId({ campaignId, userId });
-
-      // then
-      expect(response).to.be.instanceOf(CampaignParticipation);
-      expect(response.id).to.equal(campaignParticipation.id);
-    });
-
-    it('should return the non improved campaign participation found', async function () {
-      // given
-      databaseBuilder.factory.buildCampaignParticipation({
-        campaignId,
-        userId,
-        isImproved: true,
-      });
-      const campaignParticipation = databaseBuilder.factory.buildCampaignParticipation({
-        campaignId,
-        userId,
-      });
-
-      await databaseBuilder.commit();
-
-      // when
-      const response = await campaignParticipationRepository.findOneByCampaignIdAndUserId({ campaignId, userId });
-
-      // then
-      expect(response).to.be.instanceOf(CampaignParticipation);
-      expect(response.id).to.equal(campaignParticipation.id);
-    });
-
-    it('should include assessments found too', async function () {
-      // given
-      const campaignParticipation = databaseBuilder.factory.buildCampaignParticipation({
-        campaignId,
-        userId,
-      });
-      const assessment = databaseBuilder.factory.buildAssessment({ campaignParticipationId: campaignParticipation.id });
-      await databaseBuilder.commit();
-
-      // when
-      const response = await campaignParticipationRepository.findOneByCampaignIdAndUserId({ campaignId, userId });
-
-      // then
-      expect(response.assessments).to.have.lengthOf(1);
-      expect(response.assessments[0]).to.be.instanceOf(Assessment);
-      expect(response.assessments[0].id).to.equal(assessment.id);
-    });
-
-    it('should return no campaign participation', async function () {
-      // when
-      const response = await campaignParticipationRepository.findOneByCampaignIdAndUserId({ campaignId, userId });
-
-      // then
-      expect(response).to.equal(null);
     });
   });
