+++ api/tests/prescription/campaign-participation/unit/application/learner-participation-route_test.js
@@ -1,5 +1,6 @@
 import { learnerParticipationController } from '../../../../..//src/prescription/campaign-participation/application/learner-participation-controller.js';
 import * as moduleUnderTest from '../../../../../src/prescription/campaign-participation/application/learner-participation-route.js';
+import { securityPreHandlers } from '../../../../../src/shared/application/security-pre-handlers.js';
 import { expect, HttpTestServer, sinon } from '../../../../test-helper.js';
 
 describe('Unit | Application | Router | campaign-participation-router ', function () {
@@ -44,6 +45,58 @@ describe('Unit | Application | Router | campaign-participation-router ', functio
 
       // then
       expect(response.statusCode).to.equal(400);
+    });
+  });
+
+  describe('GET /api/users/{userId}/campaigns/{campaignId}/profile', function () {
+    const method = 'GET';
+
+    it('returns 200', async function () {
+      // given
+      sinon.stub(learnerParticipationController, 'getUserProfileSharedForCampaign').returns('ok');
+      sinon
+        .stub(securityPreHandlers, 'checkRequestedUserIsAuthenticatedUser')
+        .callsFake((request, h) => h.response(true));
+      const httpTestServer = new HttpTestServer();
+      await httpTestServer.register(moduleUnderTest);
+
+      const url = '/api/users/12/campaigns/34/profile';
+
+      // when
+      const result = await httpTestServer.request(method, url);
+
+      // then
+      expect(result.statusCode).to.equal(200);
+    });
+
+    it('returns 400 when userId is not a number', async function () {
+      // given
+      const httpTestServer = new HttpTestServer();
+      await httpTestServer.register(moduleUnderTest);
+
+      const userId = 'wrongId';
+      const url = `/api/users/${userId}/campaigns/34/profile`;
+
+      // when
+      const result = await httpTestServer.request(method, url);
+
+      // then
+      expect(result.statusCode).to.equal(400);
+    });
+
+    it('returns 400 when campaignId is not a number', async function () {
+      // given
+      const httpTestServer = new HttpTestServer();
+      await httpTestServer.register(moduleUnderTest);
+
+      const campaignId = 'wrongId';
+      const url = `/api/users/12/campaigns/${campaignId}/profile`;
+
+      // when
+      const result = await httpTestServer.request(method, url);
+
+      // then
+      expect(result.statusCode).to.equal(400);
     });
   });
 });
