+++ api/tests/prescription/campaign-participation/unit/infrastructure/serializers/jsonapi/shared-profile-for-campaign-serializer_test.js
@@ -1,6 +1,6 @@
-import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/shared-profile-for-campaign-serializer.js';
-import { SharedProfileForCampaign } from '../../../../../src/shared/domain/read-models/SharedProfileForCampaign.js';
-import { domainBuilder, expect, sinon } from '../../../../test-helper.js';
+import * as serializer from '../../../../../../../src/prescription/campaign-participation/infrastructure/serializers/jsonapi/shared-profile-for-campaign-serializer.js';
+import { SharedProfileForCampaign } from '../../../../../../../src/shared/domain/read-models/SharedProfileForCampaign.js';
+import { domainBuilder, expect, sinon } from '../../../../../../test-helper.js';
 
 describe('Unit | Serializer | JSONAPI | shared-profile-for-campaign-serializer', function () {
   let clock;
